Implement environment variable at services and refactor the actions files
