+++ src/data/services/fetch-api.js
@@ -18,10 +18,8 @@ function formatPreviousDate (previousDate, format) {
   return (format.replace(/mm|dd|aaaa/gi, (matched) => map[matched]))
 }
 
-// ---Aprender a criar variável de ambiente---
-
 async function getTickers (stockTickers) {
-  const response = await fetch(`https://api.polygon.io/v3/reference/tickers/${stockTickers}?apiKey=Di9sCRa_Bj2l8cpNdcSXk4E3rpAp1aFP`)
+  const response = await fetch(process.env.REACT_APP_POLYGON_API_URL + `v3/reference/tickers/${stockTickers}?apiKey=` + process.env.REACT_APP_API_KEY)
 
   return (await response.json())
 }
@@ -29,7 +27,7 @@ async function getTickers (stockTickers) {
 async function getTickerDetails (stockTickers) {
   const previousDay = formatPreviousDate(yesterdayDate, 'aaaa-mm-dd')
 
-  const response = await fetch(`https://api.polygon.io/v1/open-close/${stockTickers}/${previousDay}?apiKey=Di9sCRa_Bj2l8cpNdcSXk4E3rpAp1aFP`)
+  const response = await fetch(process.env.REACT_APP_POLYGON_API_URL + `v1/open-close/${stockTickers}/${previousDay}?apiKey=` + process.env.REACT_APP_API_KEY)
 
   return (await response.json())
 }
