--- src/data/store/actions/remove-stock.js
@@ -1,6 +1,0 @@
-export function deleteSotck (index) {
-  return {
-    type: 'ADD_DELETED_STOCK',
-    payload: index
-  }
-}
