--- src/data/store/actions/render-stock-details.js
@@ -1,6 +1,0 @@
-export function getStocksDetails (stockDetails) {
-  return {
-    type: 'RENDER_STOCK_DETAILS',
-    payload: stockDetails
-  }
-}
