--- src/data/store/actions/render-stock.js
@@ -1,6 +1,0 @@
-export function renderStocks (newStock) {
-  return {
-    type: 'RENDER_STOCK',
-    payload: newStock
-  }
-}
