--- src/data/store/actions/stock-portifolio.js
@@ -1,6 +1,0 @@
-export function addStocksToPortifolio (stock) {
-  return {
-    type: 'ADD_STOCK_TO_PORTIFOLIO',
-    payload: stock
-  }
-}
