+++ src/data/store/reducers/render-stock-details.js
@@ -1,5 +1,7 @@
+import { RENDER_STOCK_DETAILS } from '../actions'
+
 export default function reducer (state = {}, action) {
-  if (action.type === 'RENDER_STOCK_DETAILS') {
+  if (action.type === RENDER_STOCK_DETAILS) {
     return action.payload
   }
