+++ src/data/store/reducers/render-stock.js
@@ -1,5 +1,7 @@
+import { RENDER_STOCK } from '../actions/index'
+
 export default function reducer (state = {}, action) {
-  if (action.type === 'RENDER_STOCK') {
+  if (action.type === RENDER_STOCK) {
     return action.payload
   }
   return state
